add a policy class protecting threads from bad guys: class 24

+++ app/Http/Controllers/ThreadsController.php
@@ -72,6 +72,7 @@ class ThreadsController extends Controller
 
     public function destroy($chanelId, Thread $thread)
     {
+        $this->authorize('delete', $thread);
         $thread->delete();
 
         if (request()->wantsJson()) {
